refactor: Exam Icon 추가하고, QuackIcon v1, v2 분리
