+++ common/compose/src/main/kotlin/team/duckie/app/android/common/compose/ui/UserFollowingLayout.kt
@@ -25,6 +25,7 @@ import androidx.compose.ui.unit.dp
 import team.duckie.app.android.common.compose.R
 import team.duckie.app.android.common.compose.asLoose
 import team.duckie.app.android.common.compose.centerVertical
+import team.duckie.app.android.common.compose.ui.icon.v1.DefaultProfile
 import team.duckie.app.android.common.kotlin.fastFirstOrNull
 import team.duckie.app.android.common.kotlin.npe
 import team.duckie.quackquack.ui.color.QuackColor
