+++ common/compose/src/main/kotlin/team/duckie/app/android/common/compose/ui/icon/v2/DuckieIcon.kt
@@ -1,0 +1,13 @@
+/*
+ * Designed and developed by Duckie Team, 2022
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
+ */
+
+package team.duckie.app.android.common.compose.ui.icon.v2
+
+import team.duckie.app.android.common.compose.R
+import team.duckie.quackquack.material.QuackIcon
+
+val QuackIcon.Companion.Paper get() = QuackIcon(R.drawable.ic_paper_20)
